Filters registered in a filter chain were not being properly garbage collected, resulting in resources not being released correctly. This bug is directly related to filter instances not having their destroy(...) methods called.
The Solution

The patch adds an enhancement to timed execution of thred logic to better encapsulate the polling nature
of garbage collection sweeps. A call counter was created to watch the number of threads that call into
and leave the startFilterChain method; this is the ResourceMonitor interface and the related
ResourceConsumerMonitor implementation. A garbage collection service was then added to the JNDI services
suite. While adding a new service, service binding and init was refactored to be reduce code redundancy.