Add Header Value Parsing and Encapsulation
This commit adds the necessarry classes and tests to parse non-complex and complex
header values.

A header field can now be broken down one of two ways. By taking the raw header
field string or by using the HttpServletRequest to get an enumeration of header
values. The header field parser will then parse the values as a list of type
HeaderValue. These values have a convinence method for extracting a parsed copy
of the header value's quality factor.

By using the HeaderValue interface, a quality factor utility was written to then
take in a list of type HeaderValue and, by comparing quality factors, choose the
qualified header value.