Proxy Updates
Doing things to the proxy.

Discovered a little late that the origin connection logic was flawed. Servlet code expects dispatch a
certain way that was incompatible with isolating origin connection context to the request object
itself. Refining code to reuse the push decoders and moving the request/reponse logic into place.