Refactoring HeaderConstant Interface
This commit removes the getHeaderKey() method call from header constants. The reasoning was that the
toString() method provided the same functionality and that there was no reason to preserve header key
formatting as imput in the enumeration classes.

Implement Rate-Limiting Response Delegation

This commit enhances the rate limiting filter component by adding a new configuration value to allow
service implementors to turn on rate limited request response delegation. When delegation is enabled
the filter component will not return the response. Instead, the filter component will pass the
request downstream with the X-PP-Rate-Limited header added.