Fixing Default HeaderValue Choices
This commit fixes a deficiency in the parser interface that did not originally allow for
default values. This caused issues when the HeaderParser utilities would return null. To
fix this, the commit annotates with QualityFactorUtility with @Deprecated and adds a new
instance class called QualityFactorHeaderChooser which better encapsulates the logic
while adding support for default HeaderValue returns.