B-47227 - Test Repose Translation of XML with embedded JSON
A bug in SaxonEE 9.4 causes a license to be required for IdentityTransformer, even when it is not required.
This is fixed in 9.5, until then we use the XalanC Transformer in place of SaxonHE for Identity transforms.
This same procedure is used in api-checker to get around this issue.
When adding SaxonEE 9.5, need to update all sections tagged with TODO_SAXON_WORKAROUND