Updated this deproxy test
The acceptance criteria for this is not very clear at all. I have
updated the backlog item with the behavior as I understood it to be
defined.

I took out/changed the tests that will not pass based on the behavior
described by Jorge Williams for this filter. There was still some
ambiguity around the order in which things are preferred.

I understood it to treat the configuration as a priority order. Higher
ordered items are preferred to lower ordered items, with the exception
of the "preferred" type which is always the most important one.