This clobbers the ContextAdapter. This will break all the things.
It's necessary to remove dependency cycles and actually put stuff into a
core-api package that can be used by the filters.

We will likely have to replace the ContextAdapter with something
moderately more generic, but that should work out fine.

This caused ripple effects and so I started working on the PowerFilter.

It needed to be a proper spring bean, this lead to the
DelegatingFilterProxy which lead to so many other things.

This probably isn't perfect, but it gets the code to compile, and I
think is the right way to get things done. Further refactoring will
probably be necessary to move things futher to where they need to be,
rather than injecting 11 things in PowerFilter and passing them through
methods. THey can probably be injected directly.