Lots of things
Trying to get it to stand up.
The ReposeInstanceInfo I think might go away in favor of just using a
SystemModelInterrogator and adding a couple methods to provide that
information. I don't think the repose instance info is something we can
use. Of course, we also discovered that repose fires up MANY Jetty's and
that could be that before there was a single spring context for each
jetty and each one of those knew crap about itself... Either way,
insanity. Need to find a better way to do this, and when we started, we
assumed that there was only one jetty and one servlet I think, and many
filters. Ugh.