Refactor time.
Cleaned up the system model parser a bit. Made some of the lookups more
efficient for local nodes. Changed it to be a more expressive type the
Try[]. Now the function describes what's going to happen, and it's
either a Success or a Failure to parse, and I can nab exceptions that
might be thrown during the XML parsing.