Passing tests!
Had to change a few things. Servo is now a class, and main creates a new
one to run it. THis was needed for testing. The actor system is stood up
and tore down each time. Much better for testing.

Added the slf4j akka dependency so that akka log messages will go into
there

Added a couple more tests. Almost done with this guy.