Restructure validateEndpoint
Now it works off a boolean, and is just a function. It will look at the
data, and then determine if the endpoint is in the list. If there is no
AuthenticateResponse, and the endpoint validation is required, it is
unauthorized.
Refactored the tests to validate the endpoint validation method