Updated saxon-ee versions to match api-checker
Note that the saxon(-he) dependency in translation was not updated. 9.6.0.9 of that artifact is not currently available in either maven central or Rackspace's Nexus. I hope that doesn't cause issues, but if it does, then we probably should not have both dependencies in the first place. We can upload the saxon 9.6.0.9 artifact if necessary.