Updated tests for whitespace stripping
The format string will have leading and trailing whitespace stripped,
but not whitespace within the format string.

This seems to make the most sense to me, and it allows nicer formatting
of the log string. Also updated the tests to verify that the CDATA stuff
is actually being used, and that it doesn't dork with the whitespace
trimming.