Refactored this to pass through the things it needs and give us a Jetty
This jetty contains a DistDatastoreServlet, which is where most of the
work is done.

The work of the manager was just to turn things on, and it was very
complicatedly tied to the servlet initialization stuff. Instead we'll
pass it through, so that our launcher can do the initialization work,
and things are handled in spring properly.