A refactor of the servlet.
Now it doesn't try to get stuff from the fake spring servlet context
thingy and instead expects those things to be handed to it in the
constructor. This is easily done in the builder. It could be converted
to a spring context as well, if we wanted to package all the dist
datastore stuff in a spring context (which is looking smarter every
minute)