Fixing a configuration pool problem
This showed up when the configuration service and it's various
subscribing insanity was no longer happening in a very specific order.
The configuration service implementation is was not really thread safe
and the key for the confuration parser was incomplete.

This fixes that so that we don't end up getting a non-validating parser
when we supplied an XSD to enforce validation.