Slightly better concurrency management
I tried to get rid of the Synchronized in here, but the thing is rather
poorly designed, and I don't want to add the complexity of a message
queue. Basically this whole thing should be replaced by akka actors and
that way we can handle events and single threaded access to things. The
synchronization has been reduced, and I'm using concurrent HashMaps
everywhere so all views of the map will be correct.