Fixed valve launching and DistDatastore Configuration problems
Different from old repose: per-node services will be started before the
jetty comes up for a local repose instance. This is opposite behavior
from before. This means that the test removed is no longer valid.
Functionality was added to the ValveRunner to have a sanity check around
if all nodes failed to configure, exit valve.