The great datastore collapse!
Moved the 4 datastore projects into two. This also helped solve
classpath dependency hell. There was so much spaghetti between the
projects that it was a nightmare to figure out. One of the projects only
had one class in it. Now they're combined, especially since core
dependended on them all anyway. This is much less complicated to reason
about, and you have only two places to go to look up datastore stuff.

Additionally I had to move some classes into the core-service-api
package. the JAXB stuff is in there, which makes sense anyway, less
stuff needs to depend on core-lib. Also the spring properties had to go
in there, which is also perfectly acceptable.