This is why we should not use groovy
Turns out that groovy doesn't do any compile time checking for catch statements, and
so this JMX stuff has been broken since it was initially committed. We've never
been timing out on JMX errors, never been waiting. We were probably getting lucky on master.
This doesn't use any linkedin classes, just uses the groovy one, which is easier
to understand anyway. We wrapped it all in a more verbose failure log so that we'll
know what happened when JMX dies.