Split up parts of the API-validator test
Now the tests are split up. One of the tests verified mbeans after
changing the configuration, and that required restarting all the
things. The original test just verified increments in mbeans, so it only
starts repose once. The Mbean reconfiguration test starts repose fresh
each time, and then reconfigures it and restarts it.