Configure Saxon EE to use the proper ClassLoader
Saxon EE generates ByteCode for xslt transforms and stuff, so it's
stupid fast. Unfortunately, that didn't know how to use our classloader
(we would've had this problem with OSGi also). This commit gets the
DynamicLoader (which appears to be how it gets objects) and sets the
classloader to the one that would have access to the saxonica classes.
This way when it tries to load stuff, it can actually find it.