Unfortunately I can't be more specific in the failure :(
There are tests that rely on a 500 error, and the way they fail
indicates that its an internal server error, which is correct. At this
stage, without substantial research I cannot figure out how to get a 502
instead of a 500 in the right cases.

At least we won't throw a silly null pointer exception.