Fixing whatever the heck this was
Using the zinc compiler seems to have caused problems here, I don't know
why. Switching to non-incremental compilation has fixed it.

Additionally there were missing dependencies in the test scope for the
service client impl test that was built.

Finally I bumped the version of the scala maven plugin, initially in the
hopes that we could still use the incremental compilation, but no.