Lots of SSL tests!
I have to use openssl s_client to verify renegotiation. Java can't
operate at that low of a level, as best as I've been able to find. The
only requirement added here is that I must have an "openssl" on the path
to call the right commands and try renegotiation.