Here's a failing test that fails super hard.
Going to have to implement everything using sockets, because deproxy and
jetty do too much with the http response/request and I'm going to have
to implement it all by hand to ensure that the right payloads come
through. If deproxy had a "Raw" mode where I could look at the actual
text sent over the wire, I'd be good to go.