Removing this test.
I don't believe that the test is valid. The handle response method of
the handler is only called when the URI matches the limits call.

It will only get to the handleResponse logic if it has passed the rate
limits call, and if the describe limits URI pattern is present and
matches the current URI.

This means that random upstream responses will never pass through this
filter, and there's no validity for this specific test case.