Have to use the custom xjc task :(
The jaxb plugin for gradle that I found on the internet does not support
nested directories of .binding files. We'd have to specify it once for
each directory we store the binding file in. ANd it'd have to be dynamic
enough to deal with the fact that we're translating all those schema
files.