Cleanup of the "all dependencies"
These were causing insurmountable problems with the
pjl-compressingFilter. Had to disable them and come up with a better way
of getting them into the projects.

Fortunately, using the commonTestDeps on all projects is safe. The
mockrunner servlet doing that is not safe. Nor is all that java web api
insanity.