Implementation!
I also did a bit of refactoring to be able to get at the identity
response, so I could log the raw failure in case of a bad request. Along
the context lines of something Contract-y changed, and we need to know
about it. This does potentially expose authentication details, so it
might not be the best course of action.