Refactored the test pretty significantly.
Deleted some things that weren't being used. The checks were wrong,
because repose wasn't always returning the limits in the order the
checks were. I changed the test to use set theory style checks to ensure
taht all the checks we're doing are in the list of returned limits, as
well as no extra limits came out. It evidences the bug, and works FAR
more reliably. Additionally there was a bug in a check where it wasn't
even checking correctly! I'm not sure how it ever passed.