Chaning the way the DeRP filter writes the error
When writing to the PrintWriter, the message never made it to the output stream of the response body. The reason is that the PrintWriter never flushed its data.
Attempting to flush the PrintWriter and call sendError would throw an exception since either of those actions commits the response. Instead, we will just call
sendError.