Ability to support x-forwarded-for header for x-pp-groups in ip-user filter
Currently, only source IP from the calling service is matched against the CIDR
list in ip-user filter.  This is a limitation for those products that use a
load balancer in front of repose since that load balancer's ip is used instead
of the calling service's ip.  This change will mimic the behavior for x-pp-user
header population in the same filter by supporting x-forwarded-for header used
to overwrite the source ip.