Address sleeps in UriNormalizationFilterTest
These tests shared a single repose process across several test methods.
Some test methods would apply a config and wait for a config reload,
which was a 25 second sleep.

For those methods that had multiple iterations, this meant lots of
unnecessary sleeping. Since config files are static across iterations of
a test method, we really only need to applyConfigs and wait for a reload
on the first iteration.

This commit ensures the apply + wait occurs only on the first iteration
for those test methods, and replaces the sleep with a log search.